Fixed compilation warnings

+++ libecho/src/Echo.cc
@@ -29,6 +29,7 @@ void Echo::initEcho(int a_argc, char **a_argv) {
 
 void Echo::send(const std::vector<uint8_t> &buffer) {
     auto encoded = converter->encode(buffer);
+    // TODO: try to make atOnce smaller
     const size_t atOnce = 2600; /* < Temporary - this constant was created empirically.
                                      Shortly - it says how much audio data need to be pushed to stream buffer
                                      every time so there won't be pause in transmission. */
@@ -132,7 +133,7 @@ std::vector<uint8_t> Echo::receive() {
     }
 
     std::vector<uint8_t> res_bytes;
-    for (int i = 0; i < res_bits.size(); i += CHAR_BIT) {
+    for (size_t i = 0; i < res_bits.size(); i += CHAR_BIT) {
         uint8_t byte = 0;
         for (int j = 0; j < CHAR_BIT; j++) {
             byte |= (static_cast<int>(res_bits[i + j]) << j);
